requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.3
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-froala-editor==4.0.10
django-on-heroku==1.1.2
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-spectacular==0.22.1
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.6.1
MarkupSafe==2.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.1
python-decouple==3.6
python3-openid==3.2.0
pytz==2019.3
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.17.6
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.1.0
